pyproject.toml
[project]
name = "langchain-tutorials"
version = "1.0.0"
description = "LangChain 入门教程：9 个核心场景的完整示例"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.0.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-text-splitters>=1.1.0",
    "faiss-cpu>=1.7.4",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "langgraph>=1.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "langchain-mcp-adapters>=0.1.0",
    "mcp>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["demos", "helpers"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
